support loading from png via file picker
