+++ src/data/json.ts
@@ -49,8 +49,8 @@ export const saveAsJSON = async (
 export const loadFromJSON = async (appState: AppState) => {
   const blob = await fileOpen({
     description: "Excalidraw files",
-    extensions: [".json", ".excalidraw"],
-    mimeTypes: ["application/json"],
+    extensions: [".json", ".excalidraw", ".png"],
+    mimeTypes: ["application/json", "image/png"],
   });
   return loadFromBlob(blob, appState);
 };
